requirements.txt
asgiref==3.4.1
base58==2.1.0
click==8.0.3
click-aliases==1.0.1
Django==3.2.8
django-extensions==3.1.3
django-widget-tweaks==1.4.9
ecdsa==0.17.0
hdwallet==2.0.1
mnemonic==0.20
pysha3==1.0.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.9
Werkzeug==2.0.2
